fix(pins): external triggered updates of coordinate are considered
Fixes #766
